Replace inline HTML with Jinja2 templating system
This commit implements the complete Jinja2 templating migration as described in issue #18:

**Phase 1: Setup Jinja2 and Basic Structure**
- Added Jinja2 dependency to pyproject.toml
- Configured FastAPI Jinja2Templates in src/main.py
- Created templates/ and static/ directory structure
- Created base.html template with common HTML structure

**Phase 2: Extract Dashboard Templates**
- Extracted dashboard HTML from inline f-string to templates/dashboard/index.html
- Moved dashboard CSS from inline <style> to static/css/dashboard.css
- Created static/js/dashboard.js for dashboard-specific JavaScript
- Updated dashboard route to use templates.TemplateResponse instead of raw HTML

**Phase 3: Extract Chat Interface**
- Extracted chat interface HTML to templates/chat/interface.html
- Moved extensive chat CSS (~300 lines) to static/css/chat.css
- Moved complex chat JavaScript (~370 lines) to static/js/chat.js
- Updated chat route to use template rendering

**Benefits Achieved:**
- ✅ Separation of concerns: HTML, CSS, JS now in dedicated files
- ✅ Template inheritance: Both dashboard and chat extend base.html
- ✅ Static asset management: Proper file organization and serving
- ✅ Maintainability: No more inline HTML mixed with Python code
- ✅ Syntax highlighting: Proper editor support for HTML/CSS/JS
- ✅ Designer-friendly: Non-Python developers can modify templates
- ✅ Version control: Better diffs showing actual content changes

**Files Created:**
- templates/base.html (common layout)
- templates/dashboard/index.html (dashboard template)
- templates/chat/interface.html (chat interface template)
- static/css/base.css (shared styles)
- static/css/dashboard.css (dashboard styles)
- static/css/chat.css (chat interface styles)
- static/js/app.js (common JavaScript)
- static/js/dashboard.js (dashboard JavaScript)
- static/js/chat.js (chat interface JavaScript)

**Files Modified:**
- pyproject.toml (added Jinja2 dependency)
- src/main.py (removed ~750 lines of inline HTML, added template configuration)

This refactoring removes ~750 lines of inline HTML/CSS/JavaScript from Python code and creates a proper template structure that will enable easier maintenance, better collaboration, and future enhancements like dark mode support.
