+++ pyproject.toml
@@ -30,6 +30,7 @@ dependencies = [
     "typer>=0.16.0",
     "py-spy>=0.4.0",
     "httpx>=0.28.1",
+    "jinja2>=3.1.2",
 ]
 
 [project.optional-dependencies]
