+++ static/css/base.css
@@ -1,0 +1,11 @@
+/* Base styles for all pages */
+* {
+    box-sizing: border-box;
+}
+
+body {
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
+    margin: 0;
+    padding: 0;
+    color: #333;
+}
