+++ static/css/chat.css
@@ -1,0 +1,367 @@
+/* Chat interface styles */
+* { 
+    box-sizing: border-box; 
+}
+
+body { 
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif; 
+    margin: 0; 
+    padding: 0; 
+    background-color: #0f0f23; 
+    color: #ececf1;
+    height: 100vh;
+    overflow: hidden;
+}
+
+.app-container {
+    display: flex;
+    height: 100vh;
+}
+
+/* Sidebar Styles */
+.sidebar {
+    width: 260px;
+    background-color: #202123;
+    border-right: 1px solid #32343a;
+    display: flex;
+    flex-direction: column;
+    position: relative;
+}
+
+.sidebar-header {
+    padding: 12px;
+    border-bottom: 1px solid #32343a;
+}
+
+.new-chat-btn {
+    width: 100%;
+    padding: 12px;
+    background: transparent;
+    border: 1px solid #565869;
+    border-radius: 6px;
+    color: #ececf1;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 14px;
+    transition: background-color 0.2s;
+}
+
+.new-chat-btn:hover {
+    background-color: #2a2b32;
+}
+
+.threads-list {
+    flex: 1;
+    overflow-y: auto;
+    padding: 8px 0;
+}
+
+.thread-item {
+    padding: 12px;
+    margin: 2px 8px;
+    border-radius: 6px;
+    cursor: pointer;
+    font-size: 14px;
+    word-break: break-word;
+    position: relative;
+    transition: background-color 0.2s;
+}
+
+.thread-item:hover {
+    background-color: #2a2b32;
+}
+
+.thread-item.active {
+    background-color: #343541;
+}
+
+.thread-title {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+
+.user-section {
+    padding: 12px;
+    border-top: 1px solid #32343a;
+    font-size: 14px;
+    color: #8e8ea0;
+}
+
+/* Main Chat Area */
+.main-content {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    background-color: #343541;
+}
+
+.chat-header {
+    padding: 12px 20px;
+    border-bottom: 1px solid #32343a;
+    background-color: #343541;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.chat-title {
+    font-size: 18px;
+    font-weight: 600;
+    color: #ececf1;
+}
+
+.status {
+    font-size: 12px;
+    padding: 4px 8px;
+    border-radius: 4px;
+    font-weight: 500;
+}
+
+.connected { 
+    background-color: #10a37f; 
+    color: white;
+}
+
+.disconnected { 
+    background-color: #ef4444; 
+    color: white;
+}
+
+.chat-messages {
+    flex: 1;
+    overflow-y: auto;
+    padding: 20px;
+    background-color: #343541;
+}
+
+.message {
+    margin-bottom: 24px;
+    max-width: 800px;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.message-wrapper {
+    display: flex;
+    gap: 16px;
+    padding: 20px;
+    border-radius: 8px;
+}
+
+.user-message .message-wrapper {
+    background-color: #444654;
+}
+
+.ai-message .message-wrapper {
+    background-color: #343541;
+    border: 1px solid #32343a;
+}
+
+.message-role {
+    font-weight: 600;
+    font-size: 16px;
+    margin-bottom: 8px;
+    min-width: 60px;
+}
+
+.user-role { 
+    color: #19c37d; 
+}
+
+.ai-role { 
+    color: #ff6b6b; 
+}
+
+.message-content {
+    flex: 1;
+    line-height: 1.6;
+}
+
+.chat-input-container {
+    padding: 20px;
+    background-color: #343541;
+    border-top: 1px solid #32343a;
+}
+
+.chat-input {
+    max-width: 800px;
+    margin: 0 auto;
+    display: flex;
+    align-items: end;
+    gap: 12px;
+    background-color: #40414f;
+    border-radius: 12px;
+    padding: 12px;
+    border: 1px solid #565869;
+}
+
+.chat-input textarea {
+    flex: 1;
+    background: transparent;
+    border: none;
+    color: #ececf1;
+    font-size: 16px;
+    font-family: inherit;
+    resize: none;
+    outline: none;
+    min-height: 24px;
+    max-height: 200px;
+    padding: 0;
+}
+
+.chat-input button {
+    background: #19c37d;
+    border: none;
+    border-radius: 6px;
+    color: white;
+    cursor: pointer;
+    padding: 8px 12px;
+    font-size: 14px;
+    transition: background-color 0.2s;
+}
+
+.chat-input button:hover:not(:disabled) {
+    background: #0fa36b;
+}
+
+.chat-input button:disabled {
+    background: #565869;
+    cursor: not-allowed;
+}
+
+.file-button {
+    background: #40414f !important;
+    border: 1px solid #565869 !important;
+    color: #ececf1 !important;
+}
+
+.file-button:hover {
+    background: #525362 !important;
+}
+
+.typing-cursor { 
+    color: #19c37d; 
+    font-weight: bold; 
+    margin-left: 2px; 
+}
+
+@keyframes blink { 
+    0%, 50% { opacity: 1; } 
+    51%, 100% { opacity: 0; } 
+}
+
+.streaming-content { 
+    display: inline; 
+}
+
+/* Markdown styling */
+.message-content h1, .message-content h2, .message-content h3 {
+    margin: 16px 0 8px 0; 
+    color: #ececf1;
+}
+
+.message-content h1 { 
+    font-size: 1.25em; 
+}
+
+.message-content h2 { 
+    font-size: 1.1em; 
+}
+
+.message-content h3 { 
+    font-size: 1.05em; 
+}
+
+.message-content p { 
+    margin: 12px 0; 
+    line-height: 1.6; 
+}
+
+.message-content code {
+    background: #1a1b26; 
+    padding: 2px 6px; 
+    border-radius: 4px;
+    font-family: 'SF Mono', Monaco, 'Cascadia Code', 'Roboto Mono', Consolas, 'Courier New', monospace; 
+    font-size: 0.9em;
+    color: #e2e8f0;
+}
+
+.message-content pre {
+    background: #1a1b26; 
+    padding: 16px; 
+    border-radius: 8px;
+    overflow-x: auto; 
+    margin: 16px 0; 
+    border-left: 4px solid #19c37d;
+}
+
+.message-content pre code {
+    background: none; 
+    padding: 0;
+}
+
+.message-content ul, .message-content ol {
+    margin: 12px 0; 
+    padding-left: 24px;
+}
+
+.message-content li { 
+    margin: 6px 0; 
+}
+
+.message-content blockquote {
+    border-left: 4px solid #565869; 
+    margin: 16px 0; 
+    padding: 8px 16px;
+    background: #2a2b32; 
+    font-style: italic;
+}
+
+.message-content strong { 
+    font-weight: 600; 
+    color: #ececf1; 
+}
+
+.message-content em { 
+    font-style: italic; 
+}
+
+.message-content a { 
+    color: #19c37d; 
+    text-decoration: none; 
+}
+
+.message-content a:hover { 
+    text-decoration: underline; 
+}
+
+/* Loading state */
+.loading {
+    text-align: center;
+    color: #8e8ea0;
+    padding: 20px;
+    font-style: italic;
+}
+
+/* Empty state */
+.empty-chat {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    text-align: center;
+    color: #8e8ea0;
+}
+
+.empty-chat h2 {
+    margin-bottom: 8px;
+    color: #ececf1;
+}
+
+.empty-chat p {
+    margin-bottom: 24px;
+}
