+++ static/js/app.js
@@ -1,0 +1,2 @@
+// Common JavaScript for all pages
+console.log('Sample Chat App loaded');
