+++ static/js/dashboard.js
@@ -1,0 +1,4 @@
+// Dashboard-specific JavaScript
+console.log('Developer Dashboard loaded');
+
+// Add any dashboard-specific interactions here
